+++ resources/views/transaction/index.blade.php
@@ -5,10 +5,12 @@
         @csrf
         <x-laraknife.panels.standard title="{{ __('Transactions') }}" fieldset="false">
             <x-laraknife.layout.nav-tabs :info="$navTabsInfo" fieldset="true" button1Name="" button2Width1="8">
-                <x-laraknife.forms.string position="first" name="mandator" label="Mandator"
+                <x-laraknife.forms.string position="first" name="account" label="Account"
+                value="{{ $context->valueOf('account') }}" width2="3" attribute="readonly"/>
+                <x-laraknife.forms.string position="middle" name="accountAmount" label=""
+                value="{{ $context->valueOf('accountAmount') }}" width1="0" width2="1" attribute="readonly"/>
+                <x-laraknife.forms.string position="last" name="mandator" label="Mandator"
                     value="{{ $context->valueOf('mandator') }}" width2="4" attribute="readonly"/>
-                <x-laraknife.forms.string position="last" name="account" label="Account"
-                    value="{{ $context->valueOf('account') }}" width2="4" attribute="readonly"/>
                 <x-laraknife.panels.filter legend="{{ $pagination->legendText() }}">
                     <x-laraknife.forms.combobox position="first" name="transactiontype" label="Type"
                         :options="$optionsTransactiontype" class="lkn-autoupdate" width2="4" />
@@ -43,17 +45,19 @@
                             <tr>
                                 <td><x-laraknife.icons.change-record module="transaction" no="{{ $transaction->id }}" />
                                 </td>
-                                <td> {{ __($transaction->transactiontype_scope) }}</td>
-                                <td> {{ __($transaction->transactionstate_scope) }}</td>
                                 <td>{{ $transaction->name }}</td>
                                 <td>{{ $transaction->info }}</td>
                                 <td class="lkn-align-right">{{ $transaction->amount }}</td>
+                                <td> {{ __($transaction->transactiontype_scope) }}</td>
+                                <td> {{ __($transaction->transactionstate_scope) }}</td>
                                 <td>{{ $transaction->date }}</td>
                                 <td>{{ $transaction->owner }}</td>
                                 <td><x-laraknife.icons.delete-record module="transaction" no="{{ $transaction->id }}" />
                                 </td>
                             </tr>
                         @endforeach
+                        <tr><td></td><td><span class="lkn-text-tagged">{{ __('Sum')}}</span></td><td></td><td class="lkn-align-right"><span class="lkn-text-tagged">{{ $context->valueOf('sum') }}</span></td><td></td><td></td><td></td><td></td>
+                        </tr>
                     </tbody>
                 </x-laraknife.panels.sortable-table>
             </x-laraknife.layout.nav-tabs>